pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "grit"
version = "1.0"
description = "Description of my package"
dependencies = [
    'GitPython'
]

[project.scripts]
grit = "grit.main:main"
